feat: 비밀번호 암호화 기능 추가

+++ src/main/java/cat/community/NyangMunity/service/UserService.java
@@ -1,26 +1,21 @@
 package cat.community.NyangMunity.service;
 
-import cat.community.NyangMunity.config.AppConfig;
-import cat.community.NyangMunity.config.JwtTokenProvider;
-import cat.community.NyangMunity.domain.Token;
+import cat.community.NyangMunity.config.JwtTokenProvider;;
+import cat.community.NyangMunity.crypto.ScryptPasswordEncoder;
 import cat.community.NyangMunity.domain.User;
+import cat.community.NyangMunity.exception.AlreadyExistsEmailException;
 import cat.community.NyangMunity.exception.InvalidSigninInformation;
+import cat.community.NyangMunity.exception.Unauthorized;
 import cat.community.NyangMunity.repository.TokenRepository;
 import cat.community.NyangMunity.repository.UserRepository;
 import cat.community.NyangMunity.request.UserForm;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import javax.crypto.SecretKey;
-import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Date;
-import java.util.List;
 import java.util.Optional;
 
 @Slf4j
@@ -31,14 +26,15 @@ public class UserService {
     private final UserRepository userRepository;
     private final TokenRepository tokenRepository;
     private final JwtTokenProvider jwtTokenProvider;
-
-    public List<User> findAll(){
-        return userRepository.findAll();
-    }
+    private final ScryptPasswordEncoder scryptPasswordEncoder;
 
     public Long userLogin(UserForm userForm) {
-        User user = userRepository.findByEmailAndPassword(userForm.getEmail(), userForm.getPassword())
+        User user = userRepository.findByEmail(userForm.getEmail())
                 .orElseThrow(InvalidSigninInformation::new);
+
+        if(!scryptPasswordEncoder.matches(userForm.getPassword(), user.getPassword())) {
+            throw new InvalidSigninInformation();
+        }
 
         if(!user.getTokens().isEmpty()) {
             tokenRepository.deleteByUserId(user.getId());
@@ -50,16 +46,20 @@ public class UserService {
     }
 
     public void register(UserForm userForm) {
-        User user = User.builder()
-                .email(userForm.getEmail())
-                .password(userForm.getPassword())
-                .nickname(userForm.getNickname())
-                .birthday(userForm.getBirthday() != null && !userForm.getBirthday().isEmpty()
-                        ? LocalDate.parse(userForm.getBirthday(), DateTimeFormatter.ofPattern("yyyy-MM-dd")) : null)
-                .createDate(LocalDateTime.now())
-                .build();
-
-        userRepository.save(user);
+        Optional<User> userCheck = userRepository.findByEmail(userForm.getEmail());
+        if(userCheck.isPresent()){
+            throw new AlreadyExistsEmailException();
+        } else {
+            User user = User.builder()
+                    .email(userForm.getEmail())
+                    .password(scryptPasswordEncoder.encrypt(userForm.getPassword()))
+                    .nickname(userForm.getNickname())
+                    .birthday(userForm.getBirthday() != null && !userForm.getBirthday().isEmpty()
+                            ? LocalDate.parse(userForm.getBirthday(), DateTimeFormatter.ofPattern("yyyy-MM-dd")) : null)
+                    .createDate(LocalDateTime.now())
+                    .build();
+            userRepository.save(user);
+        }
     }
 
     public User userCheck(Long userId) {
@@ -69,6 +69,13 @@ public class UserService {
 
     public void userLogout(Long userId) {
         tokenRepository.deleteByUserId(userId);
+    }
+
+    public User userInfo(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(Unauthorized::new);
+
+        return user;
     }
 }
